Fixed a big integer conversion bug, updated to version 0.3.1

+++ src/src/gamrcorps/convex/Conv.java
@@ -285,8 +285,8 @@ public class Conv {
     }
 
 	public static Object simplify(double d) {
-		if (d == (int) d){
-			return (int) d;
+		if (d == (long) d){
+			return (long) d;
 		}
 		return d;
 	}
